requirements.txt
watchdog ==5.0.3   
paho-mqtt==1.6.1
pandas ==2.2.3
python-dotenv